Fixed incorrect use of strip()
strip() was used to remove end of a filename, which worked most of the time
however it would also remove all those characters from elsewhere in the
filename. Thus bug manifest when I passed filenames in the form ./dir/file
when it stripped the leading period, among other characters.

I swapped it out for replace() which works great