rearranged import errors to avoid throwing the same warning due to recursive imports (mucor imports output imports from mucor)
created a wrapper function for pandas sort to make it agnostic of which pandas version is used. Pandas version >= 0.17.0 implements a new/different sort function.

implemented functionality to skip tabix and xlsx features if the relevant modules are missing. Although checks were in place, no action was taken for missing modules.